requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.3
openai==1.35.14
pinecone-client==3.2.2
google-generativeai==0.3.0
python-dotenv==1.0.0
feedparser==6.0.10
APScheduler==3.10.4
numpy==1.24.3
pydantic==2.5.0
aiohttp==3.9.1
requests==2.31.0
